Fix for fetch.js fetchEvents function
The former version omits the JSON parse exceptions. Since the type of "params.qstring.events" still a string after the parse attempt, the if block doesn't get executed, and the function exits silently. Naturally, the system doesn't generate a response for requests that contain bad "events" querystring, and leads to a timeout.