fix: Map picker render error on drawer
Map picker does not update its size when the drawer opens. Its size
still remains when the drawer is mounted.

Add height prop to map picker so that it uses static units.

Add invalidateSize() method to map picker which dispatches a resize
event that is caught by the leaflet map.